@@ -13,36 +13,34 @@
     "publishnext": "npm publish --tag next",
     "prepublishOnly": "yarn build",
     "postpublish": "git push --tags",
-    "test": "eslint src examples"
+    "test": "eslint src examples",
+    "test-webpack": "jest"
   },
   "repository": "nozzle/react-static",
   "devDependencies": {
-    "babel-eslint": "^7.2.3",
-    "eslint": "^4.1.1",
-    "eslint-config-react-tools": "^1.0.10",
-    "eslint-plugin-class-property": "^1.0.6",
-    "eslint-plugin-flowtype": "^2.34.1",
-    "eslint-plugin-import": "^2.6.0",
-    "eslint-plugin-jsx-a11y": "^5.1.1",
-    "eslint-plugin-react": "^7.1.0",
+    "eslint": "^4.3.0",
+    "eslint-config-react-tools": "^1.0.6",
+    "jest": "^21.2.1",
     "rimraf": "^2.6.2"
   },
   "dependencies": {
-    "@babel/cli": "^7.0.0-beta.32",
-    "@babel/core": "^7.0.0-beta.32",
-    "@babel/plugin-transform-runtime": "^7.0.0-beta.32",
-    "@babel/preset-env": "^7.0.0-beta.32",
-    "@babel/preset-react": "^7.0.0-beta.32",
-    "@babel/preset-stage-0": "^7.0.0-beta.32",
-    "@babel/register": "^7.0.0-beta.32",
-    "@babel/runtime": "^7.0.0-beta.32",
-    "@std/esm": "^0.14.0",
     "@types/react": "^16.0.18",
     "@types/react-helmet": "^5.0.3",
     "@types/react-router-dom": "^4.2.0",
     "autoprefixer": "^7.1.5",
     "axios": "^0.16.2",
-    "babel-loader": "^8.0.0-beta.0",
+    "babel-cli": "^6.26.0",
+    "babel-loader": "^7.1.2",
+    "babel-plugin-transform-class-properties": "^6.24.1",
+    "babel-plugin-transform-runtime": "^6.23.0",
+    "babel-preset-env": "^1.6.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-latest": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-0": "^6.24.1",
+    "babel-preset-stage-1": "^6.24.1",
+    "babel-preset-stage-3": "^6.24.1",
+    "babel-runtime": "^6.26.0",
     "case-sensitive-paths-webpack-plugin": "^2.1.1",
     "circular-dependency-plugin": "^4.2.1",
     "cors": "^2.8.4",
@@ -65,7 +63,7 @@
     "react-dev-utils": "^4.0.1",
     "react-dom": "^16",
     "react-helmet": "^5.2.0",
-    "react-hot-loader": "^3.1.2",
+    "react-hot-loader": "^3.0.0-beta.7",
     "react-router-dom": "^4.2.2",
     "script-ext-html-webpack-plugin": "^1.8.7",
     "style-loader": "^0.19.0",
